added remove-admin(mapping and function) and refactorized
