+++ src/main/java/com/bureaucracyhacks/refactorip/controllers/AdminController.java
@@ -1,5 +1,6 @@
 package com.bureaucracyhacks.refactorip.controllers;
 
+import com.bureaucracyhacks.refactorip.services.AdminService;
 import com.bureaucracyhacks.refactorip.services.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -15,12 +16,28 @@ public class AdminController {
     @Autowired
     private UserService userService;
 
-    @PostMapping("/makeadmin")
+    @Autowired
+    private AdminService adminService;
+
+    @PostMapping("/make-admin")
     public ResponseEntity<?> makeAdmin(@RequestParam String username) {
         System.out.println("User is now an admin");
         if (userService.isUsernameTaken(username)) {
-            userService.makeAdmin(username);
+            adminService.makeAdmin(username);
             return new ResponseEntity<>("User is now admin!", HttpStatus.OK);
+        }
+        else
+        {
+            return new ResponseEntity<>("User does not exist!", HttpStatus.BAD_REQUEST);
+        }
+    }
+
+    @PostMapping("/remove-admin")
+    public ResponseEntity<?> removeAdmin(@RequestParam String username) {
+        System.out.println("User is no longer an admin");
+        if (userService.isUsernameTaken(username)) {
+            adminService.removeAdmin(username);
+            return new ResponseEntity<>("User is no longer admin!", HttpStatus.OK);
         }
         else
         {
